encoding/api-encoding-names.html

<!DOCTYPE html>
<title>Encoding API: Encoding names</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/shared.js"></script>
<script>

test(function() {

    var encodings = [
        { label: 'utf-8', encoding: 'utf-8' },
        { label: 'utf-16', encoding: 'utf-16le' },
        { label: 'utf-16le', encoding: 'utf-16le' },
        { label: 'utf-16be', encoding: 'utf-16be' },
        { label: 'ascii', encoding: 'windows-1252' },
        { label: 'iso-8859-1', encoding: 'windows-1252' }
    ];

    // encoding-labels.html tests the full set of names/labels; this test just
    // exercises some common cases and case-insensitivity.

    encodings.forEach(function(t) {
        assert_equals(new TextDecoder(t.label.toLowerCase()).encoding, t.encoding);
        assert_equals(new TextDecoder(t.label.toUpperCase()).encoding, t.encoding);
    });

}, 'Encoding labels are case-insensitive');

</script>
